Centralize common exchange operations on abstract exchange
